requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
streamlit==1.45.1
sqlalchemy==2.0.41
passlib[bcrypt]==1.7.4
python-jose==3.5.0
pydantic==2.11.7
requests==2.32.4
python-dotenv==1.1.0
psycopg2-binary==2.9.10
pydantic[email]==2.11.7
